fix bug

+++ html/script/manage.js
@@ -63,7 +63,10 @@ $(document).ready(function() {
                     row.find('input.input-text').prop('readonly', readonly);
                     row.find('.btn-delete').click((function(countryId, listIndex) {
                         return function() {
-                            alert('确定删除吗？');
+                            var result = confirm('确定删除吗？');
+                            if (!result) {
+                                return;
+                            }
                             var data = {
                                 countryId: countryId,
                                 listIndex: listIndex - 1,
@@ -246,6 +249,9 @@ $(document).ready(function() {
             case 'system-set':
                 initSystemSet();
                 break;
+            case 'flush-btn':
+                flush();
+                break;
             default:
                 break;
         }
@@ -271,7 +277,7 @@ $(document).ready(function() {
                 return function() {
                     var province = getProvinceData();
                     if (province[i].provinceId in usedProvince) {
-                        alert('存在省份引用，无法删除，请检查！');
+                        confirm('存在省份引用，无法删除，请检查！');
                         return;
                     }
                     province.splice(i, 1);
@@ -470,7 +476,10 @@ $(document).ready(function() {
                     row.find('select.input-select').prop('disabled', readonly);
                     row.find('.btn-school-delete').click((function(schoolId, listIndex) {
                         return function() {
-                            alert('确定删除吗？');
+                            var result = confirm('确定删除吗？');
+                            if (!result) {
+                                return;
+                            }
                             var countryListIndex = $('#school-page-country-select').val();
                             var data = {
                                 countryListIndex: countryListIndex - 1,
@@ -753,7 +762,10 @@ $(document).ready(function() {
 
                     row.find('.btn-item-delete').click((function(itemId, listIndex) {
                         return function() {
-                            alert('确定删除吗？');
+                            var result = confirm('确定删除吗？');
+                            if (!result) {
+                                return;
+                            }
                             var countryListIndex = $('#item-page-country-select').val();
                             var schoolListIndex = $('#item-page-school-select').val();
                             var data = {
@@ -990,7 +1002,10 @@ $(document).ready(function() {
                     row.find('input.input-text').prop('readonly', readonly);
                     row.find('.btn-delete').click((function(userId, listIndex) {
                         return function() {
-                            alert('确定删除吗？')
+                            var result = confirm('确定删除吗？')
+                            if (!result) {
+                                return;
+                            }
                             var data = {
                                 userId: userId,
                                 listIndex: listIndex - 1,
@@ -1061,7 +1076,7 @@ $(document).ready(function() {
 
     }
 
-    function initSystemSet(){
+    function initSystemSet() {
         $('#add-type-btn').prop('disabled', true);
         var typeSwitch = $('#system-switch input[type="checkbox"]');
         typeSwitch.prop('checked', false);
@@ -1069,6 +1084,7 @@ $(document).ready(function() {
             var readonly = !$(this).prop('checked');
             $('#add-type-btn').prop('disabled', readonly);
             $('#user-level-input').prop('readonly', readonly);
+            $('.input-text').prop('readonly', readonly);
         });
         fetchSystemData();
     }
@@ -1112,7 +1128,10 @@ $(document).ready(function() {
                 row.find('input.input-text').prop('readonly', readonly);
                 row.find('.btn-delete').off('click').click((function(listIndex) {
                     return function() {
-                        alert('确定删除吗？');
+                        var result = confirm('确定删除吗？');
+                        if (!result) {
+                            return;
+                        }
                         $.ajax({
                             url: "/system/delete",
                             type: "DELETE",
@@ -1126,7 +1145,7 @@ $(document).ready(function() {
                                 },
                                 400: function(jqXHR) {
                                     var response = JSON.parse(jqXHR.responseText);
-                                    alert(response.usedSchool);
+                                    confirm(response.usedSchool);
                                 }
                             }
                         });
@@ -1169,6 +1188,16 @@ $(document).ready(function() {
 
     function enableSidebarButton(){
         $(".button-list-button").prop("disabled", false);
+    }
+
+    function flush() {
+        var result = confirm('确定清除缓存吗？');
+        if (!result) {
+            return;
+        }
+        $.get("/flush", function(response) {
+            $('#manage-country').click();
+        });
     }
 
     fetchCountryData();
